+++ techpack/display/msm/samsung/ANA38401_AMSA05RB06/ss_dsi_interpolation_ANA38401_AMSA05RB06.c
@@ -285,10 +285,10 @@ int table_parsing_data_ANA38401_AMSA05RB06(struct samsung_display_driver_data *v
 	}__packed;
 
 	struct hbm_table_format *hbm_table = (struct hbm_table_format *)table_hbm_br_info;
-	int hbm_step = sizeof(table_hbm_br_info) / sizeof(struct hbm_table_format);
+	int hbm_step = (int)sizeof(table_hbm_br_info) / (int)sizeof(struct hbm_table_format);
 
 	struct normal_table_format *normal_table = (struct normal_table_format *)table_normal_br_info;
-	int normal_step = sizeof(table_normal_br_info) / sizeof(struct normal_table_format);
+	int normal_step = (int)sizeof(table_normal_br_info) / (int)sizeof(struct normal_table_format);
 
 	int input_table_size = sizeof(table_hbm_br_info) + sizeof(table_normal_br_info);
 	//unsigned char *dst = vdd->panel_br_info.br_data_raw;
